requirements.txt
hikari-lightbulb
peewee==3.14.8